pipeline/requirements.txt
boto3==1.38.13
botocore==1.38.13
jmespath==1.0.1
numpy==2.2.5
pandas==2.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
s3transfer==0.12.0
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
